Add ObservableArray.batchedChanges.
This implements tracking of elements that change position for
all observable arrays whose elements are hashable. Reported changes can
be plugged directly into a UITableView or UICollectionView.

This is an improvement over ArrayChange.deletedIndices/insertedIndices, but
it only makes sense when elements aren’t duplicates and they have stable identities.