Reconsider batchedChanges; fix moves.
To fix timing issues, the current state of the array needs to be available. So you cannot get source for batched changes on its own; you need a special observable array to go with it.

This is due to a design bug — ArrayChange should include references to removed elements. Once that is fixed, we’ll be able to derive batchedChanges directly from each ArrayChange, so it can be an automatic member in each ObservableArray.